--- a/Cargo.toml
+++ b/Cargo.toml
@@ -23,8 +23,8 @@ members = [ "derive" ]
 
 [dependencies]
 byteorder = "1"
-nom = "3"
-zip = { version = "0.5", optional = true }  # NOTICE: also in dev-dependencies
+py_literal = "0.4"
+zip = { version = "0.6", optional = true }  # NOTICE: also in dev-dependencies
 
 # NOTE: public dependencies, so make sure the doc links in lib.rs are kept in sync
 num-complex = { version = "0.4", optional = true }
@@ -41,10 +41,10 @@ default-features = false
 # breaking semver bumps very frequently.
 #
 # Also, sprs has an ndarray dependency that might not be the most recent.
-ndarray = { version = "0.14" }
-sprs = { version = "0.10", default-features = false }
+ndarray = { version = "0.15" }
+sprs = { version = "0.11", default-features = false }
 bencher = { version = "0.1" }
-zip = { version = "0.5" }  # NOTICE: also in dependencies
+zip = { version = "0.6" }  # NOTICE: also in dependencies
 
 [features]
 default = []