--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,7 @@ members = [ "derive" ]
 byteorder = "1"
 py_literal = "0.4"
 zip = { version = "0.6", optional = true }  # NOTICE: also in dev-dependencies
+num-bigint = "0.4"
 
 # NOTE: public dependencies, so make sure the doc links in lib.rs are kept in sync
 num-complex = { version = "0.4", optional = true }