--- a/src/header.rs
+++ b/src/header.rs
@@ -1,8 +1,7 @@
 
-use std::collections::HashMap;
 use std::io;
 
-use nom::IResult;
+pub use py_literal::Value;
 use byteorder::{LittleEndian, ReadBytesExt};
 
 use crate::type_str::TypeStr;
@@ -88,7 +87,7 @@ impl DType {
     // not part of stable API, but needed by the serialize_array test
     #[doc(hidden)]
     pub fn parse(source: &str) -> io::Result<Self> {
-        let descr = parse_header_text_to_io_result(source.as_bytes())?;
+        let descr = parse_header_text_to_io_result(&[source.as_bytes(), b"\n"].concat())?;
         Self::from_descr(&descr)
     }
 
@@ -124,6 +123,7 @@ fn convert_list_to_record_fields(values: &[Value]) -> io::Result<Vec<Field>> {
     values.iter()
         .map(|value| match *value {
             Value::List(ref tuple) => convert_tuple_to_record_field(tuple),
+            Value::Tuple(ref tuple) => convert_tuple_to_record_field(tuple),
             _ => Err(invalid_data("list must contain list or tuple"))
         })
         .collect()
@@ -154,15 +154,18 @@ fn convert_tuple_to_record_field(tuple: &[Value]) -> io::Result<Field> {
 fn convert_value_to_field_shape(field: &Value) -> io::Result<Vec<u64>> {
     if let Value::List(ref lengths) = *field {
         lengths.iter().map(convert_value_to_positive_integer).collect()
+    } else if let Value::Tuple(ref lengths) = *field {
+        lengths.iter().map(convert_value_to_positive_integer).collect()
     } else {
         Err(invalid_data("shape must be list or tuple"))
     }
 }
 
 fn convert_value_to_positive_integer(number: &Value) -> io::Result<u64> {
-    if let Value::Integer(number) = *number {
-        if number > 0 {
-            Ok(number as u64)
+    if let Value::Integer(number) = number {
+        let parts = number.to_u64_digits().1;
+        if !parts.is_empty() && parts[0] > 0 {
+            Ok(parts[0])
         } else {
             Err(invalid_data("number must be positive"))
         }
@@ -183,15 +186,18 @@ fn extract_full_array_shape(mut dtype: &DType) -> (Vec<u64>, &DType) {
 pub(crate) fn convert_value_to_shape(field: &Value) -> io::Result<Vec<u64>> {
     if let Value::List(ref lengths) = *field {
         lengths.iter().map(convert_value_to_shape_integer).collect()
+    } else if let Value::Tuple(ref lengths) = *field {
+        lengths.iter().map(convert_value_to_shape_integer).collect()
     } else {
         Err(invalid_data("shape must be list or tuple"))
     }
 }
 
 pub fn convert_value_to_shape_integer(number: &Value) -> io::Result<u64> {
-    if let Value::Integer(number) = *number {
-        if number >= 0 {
-            Ok(number as u64)
+    if let Value::Integer(number) = number {
+        let parts = number.to_u64_digits().1;
+        if !parts.is_empty() && parts[0] > 0 {
+            Ok(parts[0])
         } else {
             Err(invalid_data("shape integer cannot be negative"))
         }
@@ -211,15 +217,6 @@ fn invalid_data(message: impl ToString) -> io::Error {
     io::Error::new(io::ErrorKind::InvalidData, message.to_string())
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Value {
-    String(String),
-    Integer(i64),
-    Bool(bool),
-    List(Vec<Value>),
-    Map(HashMap<String, Value>),
-}
-
 pub(crate) fn read_header(r: &mut dyn io::Read) -> io::Result<Value> {
     let PreHeader { version_props, header_size } = read_pre_header(r)?;
 
@@ -232,18 +229,17 @@ pub(crate) fn read_header(r: &mut dyn io::Read) -> io::Result<Value> {
 }
 
 fn parse_header_text_to_io_result(bytes: &[u8]) -> io::Result<Value> {
-    match parser::value(bytes) {
-        IResult::Done(remainder, header) => {
-            if !remainder.iter().all(|b| b" \t\n\r\0".contains(b)) {
-                return Err(invalid_data(format_args!("unexpected trailing data in header: {:?}", remainder)));
-            }
-            Ok(header)
+    let without_newline = match bytes.split_last() {
+        Some((&b'\n', rest)) => rest,
+        Some(_) | None => return Err(invalid_data("missing newline")),
+    };
+    let as_str = std::str::from_utf8(without_newline).map_err(|_| invalid_data("could not parse utf-8"))?;
+    match as_str.parse() {
+        Ok(value) => {
+            Ok(value)
         },
-        IResult::Incomplete(needed) => {
-            Err(invalid_data(format_args!("could not parse Python expression: {:?}", needed)))
-        },
-        IResult::Error(err) => {
-            Err(invalid_data(format_args!("could not parse Python expression: {:?}", err)))
+        Err(err) => {
+            Err(invalid_data(format_args!("could not parse Python expression: {:?}", err.to_string())))
         },
     }
 }
@@ -326,80 +322,6 @@ pub(crate) fn get_minimal_version(required_props: VersionProps) -> (u8, u8) {
     } else {
         (1, 0)
     }
-}
-
-mod parser {
-    use super::Value;
-    use nom::*;
-
-    named!(pub value<Value>, alt!(integer | boolean | string | list | map));
-
-    named!(pub integer<Value>,
-        map!(
-            map_res!(
-                map_res!(
-                    ws!(digit),
-                    ::std::str::from_utf8
-                ),
-                ::std::str::FromStr::from_str
-            ),
-            Value::Integer
-        )
-    );
-
-    named!(pub boolean<Value>,
-        ws!(alt!(
-            tag!("True") => { |_| Value::Bool(true) } |
-            tag!("False") => { |_| Value::Bool(false) }
-        ))
-    );
-
-    named!(pub string<Value>,
-        map!(
-            map!(
-                map_res!(
-                    ws!(alt!(
-                        delimited!(tag!("\""),
-                            is_not_s!("\""),
-                            tag!("\"")) |
-                        delimited!(tag!("\'"),
-                            is_not_s!("\'"),
-                            tag!("\'"))
-                        )),
-                    ::std::str::from_utf8
-                ),
-                |s: &str| s.to_string()
-            ),
-            Value::String
-        )
-    );
-
-    named!(pub list<Value>,
-        map!(
-            ws!(alt!(
-                delimited!(tag!("["),
-                    terminated!(separated_list!(tag!(","), value), alt!(tag!(",") | tag!(""))),
-                    tag!("]")) |
-                delimited!(tag!("("),
-                    terminated!(separated_list!(tag!(","), value), alt!(tag!(",") | tag!(""))),
-                    tag!(")"))
-            )),
-            Value::List
-        )
-    );
-
-    named!(pub map<Value>,
-        map!(
-            ws!(
-                delimited!(tag!("{"),
-                    terminated!(separated_list!(tag!(","),
-                        separated_pair!(map_opt!(string, |it| match it { Value::String(s) => Some(s), _ => None }), tag!(":"), value)
-                    ), alt!(tag!(",") | tag!(""))),
-                    tag!("}"))
-            ),
-            |v: Vec<_>| Value::Map(v.into_iter().collect())
-        )
-    );
 }
 
 
@@ -539,7 +461,7 @@ mod tests {
 
     #[test]
     fn errors_on_value_variants_that_cannot_be_converted() {
-        let no_dtype = Value::Bool(false);
+        let no_dtype = Value::Boolean(false);
         assert!(DType::from_descr(&no_dtype).is_err());
     }
 
@@ -551,7 +473,7 @@ mod tests {
 
     #[test]
     fn errors_when_record_list_entry_contains_too_few_items() {
-        let faulty_list = parse("[('a')]");
+        let faulty_list = parse("[('a',)]");
         assert!(DType::from_descr(&faulty_list).is_err());
     }
 
@@ -585,8 +507,6 @@ mod tests {
     }
 
     fn parse(source: &str) -> Value {
-        parser::value(source.as_bytes())
-            .to_result()
-            .expect("could not parse Python expression")
+        source.parse().expect("could not parse Python expression")
     }
 }