Uses [`py_literal`](https://crates.io/crates/py_literal) for parsing, removing the dependency on nom 3. Fixes #56.