@ExpHP 2023-01-29
Thanks, I like this change in spirit.  It has always bothered me how the NPY spec specifies the header as a "Python literal expression of a dictionary," since a truly correct parser for this would be a fair amount of work.

I'll take a more detailed look when I have more time (hopefully tonight).  I'll especially want to look at the number of recursive dependencies; the handrolled parser wasn't great, but if there's one good thing about `nom`, it's the lack of recursive dependencies.