@ExpHP 2023-01-30
Okay, here's an analysis of the dependency changes:

**Build times for `--no-default-features`**
(i7-12700K, 8x2 threads at 3.61GHz)
| | Before | After |
| --- | --- | --- |
| Debug | 1.727s | 4.929s |
| Release | 2.451s | 6.039s |

```
$ rm -rf target && time cargo build --no-default-features
    Updating crates.io index
   Compiling libc v0.2.131
   Compiling byteorder v1.4.3
   Compiling memchr v1.0.2
   Compiling nom v3.2.1
   Compiling npyz v0.7.3 (/home/exp/asd/clone/npyz)
    Finished dev [unoptimized + debuginfo] target(s) in 1.70s
```

```
$ rm -rf target && time cargo build --no-default-features
    Updating crates.io index
   Compiling proc-macro2 v1.0.43
   Compiling autocfg v1.1.0
   Compiling unicode-ident v1.0.3
   Compiling quote v1.0.21
   Compiling syn v1.0.99
   Compiling thiserror v1.0.38
   Compiling ucd-trie v0.1.5
   Compiling once_cell v1.17.0
   Compiling byteorder v1.4.3
   Compiling num-traits v0.2.15
   Compiling num-integer v0.1.45
   Compiling num-bigint v0.4.3
   Compiling num-complex v0.4.2
   Compiling thiserror-impl v1.0.38
   Compiling pest v2.5.4
   Compiling pest_meta v2.5.4
   Compiling pest_generator v2.5.4
   Compiling pest_derive v2.5.4
   Compiling py_literal v0.4.0
   Compiling npyz v0.7.3 (/home/exp/asd/clone/npyz)
    Finished dev [unoptimized + debuginfo] target(s) in 4.90s
```

The Cargo.lock also grows approximately by a factor of 2.

Looking at the dependencies, I can group them as follows:

- `syn`/`quote`/`proc_macro2`/`unicode-ident`.  These days it's tough to keep proc macros out of most crates and it's frankly a miracle that npyz has been able to keep these deps optional to this day.
- `thiserror`.  Though never truly necessary, I think `thiserror` is a reasonably cheap dependency to have in any dep tree that already uses proc macros, and I generally encourage its use both in library and binary code.
- Various `num-*` crates.  I'm not a fan of the the num traits, but most users of npyz will also likely be using `ndarray` which also depends on these.
- `pest`.  I don't fault py-literal for using a parser generator, and this looks pretty popular and well-maintained.
- `py-literal` itself.  It's maintained by bluss and Jim Turner, two pillars of the rust community.

---

While I do have a bit of pride in npyz's low dependency count at the moment, the current parser is just plain *incorrect.*  All of these dependencies seem acceptable to have in the sort of projects where `npyz` is most likely to be used (high-level or numerical computing code), so I do not consider the additional deps to be a blocker.

I will review the changes in detail shortly.