@sudo-carson 2023-02-01
Addressed your comments in https://github.com/ExpHP/npyz/pull/60/commits/e2dddbce5239eb66dbebc1e95e5b12e776a9c397 and https://github.com/ExpHP/npyz/pull/60/commits/7be763b0504a93e7f0038482a3fb1ecdd5c4716f, please let me know if there's anything else I should change.