@ExpHP 2023-02-01
npyz 0.7.4 has been released, including this change and some fixes for structured arrays (zero-size arrays and funny characters in member names)